Fix static files samples and docs
