+++ aspnet/fundamentals/static-files/sample/Program.cs
@@ -1,21 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.IO;
+using System.Reflection;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 
-namespace noAuth
+namespace StaticFiles
 {
     public class Program
     {
         public static void Main(string[] args)
         {
+            var builder = new ConfigurationBuilder()
+                .AddCommandLine(args)
+                .AddEnvironmentVariables();
+
             var host = new WebHostBuilder()
                 .UseKestrel()
                 .UseContentRoot(Directory.GetCurrentDirectory())
                 .UseIISIntegration()
-                .UseStartup<Startup>()
+                .UseStartup(typeof(Program).GetTypeInfo().Assembly.GetName().Name)
+                .UseConfiguration(builder.Build())
                 .Build();
 
             host.Run();
