+++ aspnet/fundamentals/static-files/sample/StartupEmpty.cs
@@ -1,22 +1,22 @@
-﻿//#define StartupEmpty
-#if StartupEmpty
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace empty
+namespace StaticFiles
 {
     public class StartupEmpty
     {
+        // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
         }
 
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        // >Configure
         public void Configure(IApplicationBuilder app)
         {
             app.UseDefaultFiles();
             app.UseStaticFiles();
         }
+        // <Configure
     }
 }
-
-#endif
