+++ aspnet/fundamentals/static-files/sample/StartupTwoStaticFiles.cs
@@ -8,34 +8,24 @@ using Microsoft.Extensions.Logging;
 
 namespace StaticFiles
 {
-    public class StartupBrowse
+    public class StartupTwoStaticFiles
     {
         // This method gets called by the runtime. Use this method to add services to the container.
-        // >Services
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddDirectoryBrowser();
         }
-        // <Services
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         // >Configure
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
-            app.UseStaticFiles(); // For the wwwroot folder
+            app.UseStaticFiles();
 
             app.UseStaticFiles(new StaticFileOptions()
             {
                 FileProvider = new PhysicalFileProvider(
-                    Path.Combine(Directory.GetCurrentDirectory(), @"wwwroot\images")),
-                RequestPath = new PathString("/MyImages")
-            });
-
-            app.UseDirectoryBrowser(new DirectoryBrowserOptions()
-            {
-                FileProvider = new PhysicalFileProvider(
-                    Path.Combine(Directory.GetCurrentDirectory(), @"wwwroot\images")),
-                RequestPath = new PathString("/MyImages")
+                    Path.Combine(Directory.GetCurrentDirectory(), @"MyStaticFiles")),
+                RequestPath = new PathString("/StaticFiles")
             });
         }
         // <Configure
