--- aspnet/fundamentals/static-files/sample/Views/_ViewImports.cshtml
@@ -1,2 +1,0 @@
-﻿@using noAuth
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
