--- aspnet/fundamentals/static-files/sample/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "asp.net",
-  "version": "0.0.0",
-  "private": true,
-  "devDependencies": {
-    "gulp": "3.8.11",
-    "gulp-concat": "2.5.2",
-    "gulp-cssmin": "0.1.7",
-    "gulp-uglify": "1.2.0",
-    "rimraf": "2.2.8"
-  }
-}
